+++ tests/testthat/test-pact_list.R
@@ -3,15 +3,16 @@
 ## Tests ----
 test_that("pact_list outputs as expected", {
   skip_if(Sys.getenv("FIGSHARE_TOKEN") == "")
+  skip_on_cran()
+  skip_if_offline()
 
   pact_client <- pact_client_set()
 
   ## Get lists ----
-  list_all <- pact_list(pact_client)
-  list_data <- pact_list_data(pact_client)
+  list_oxford <- pact_figshare_list(pact_client)
+  list_private <- pact_figshare_list(pact_client)
 
-  expect_s3_class(list_all, "data.frame")
-  expect_s3_class(list_data, "data.frame")
-  expect_equal(unique(list_all$group_id), 53043, ignore_attr = TRUE)
-  expect_equal(unique(list_data$group_id), 53043, ignore_attr = TRUE)
+  expect_s3_class(list_oxford, "data.frame")
+  expect_s3_class(list_private, "data.frame")
+  expect_equal(unique(list_oxford$group_id), 53043, ignore_attr = TRUE)
 })